Set the mode for /usr/bin/groupmems to 2750
As the main use-case for `groupmems` is to allow members of the
`groups` group to manage their own groups, the group needs access to
the executable.
Adapt PKGBUILD to use a mode of 750 for `groupmems` so that pacman does
not emit a warning about the discrepancy between the mode of the file
when installed and when later on modified by tmpfiles.d.

Signed-off-by: David Runge <dvzrv@archlinux.org>